Codon Table changes
-We changed the constructor for CodonTable. We only altered the
constructor that takes arguments. Now the constructor takes 4 arguments:
codonTableId, model, split, and groupList. Both split and groupList have
default values (true and an empty vector). Because of this we had to
create a static vector to be able to pass an empty vector as a default
argument. We also had to alter the functions affected by the
constructor, such as those concerning the creation of the singleton
instance and those that create the codon table from main and in Genome.

-We implemented two new constant static vectors that tell what the
default Amino Acids are for each codon table. This is necessary for when
the user does not specify a group list. We chose to make these variables
instead of doing a if/else structure for code readability and to prevent
bugs. The difference between the two variables is if the amino acids are
split or not. Again, this could be taken care of by an if/else block,
but readability and lack of bugs led to the decision for a second
variable.

-We are in the process of rewriting the setupCodonTable to fit our new
indexing model. Currently the AAMap and AAToNumCodonsMap are
re-implemented. In order to get the number of codons for each AA in our
group list, we needed to introduce a new map that has the complete amino
acid list with a given set of indices to match our table that holds the
number of codons for each AA. This variable is called fullAAMap.

We created a large table called codonTableListing, which is a triply
indexed vector (codon table index, AA index, Codon). This stores how
each codon is encoded for every table. Now, we have to figure out what
to do with split AAs, and we have to make a mapping from the groupList
to the new codonTable vector.

TODO:

- Reimplement all private variables and setupCodonTable
- Error check user entered group List
- Figure out how RFP will work with Codon Table