CodonTable updates
- Finished setting up private variables and mapping functions
- Set up R wrappers to compile for C++ side, but they could be
inaccurate at the moment
- We tried a single run which crashed on process sequence. codonToIndex
failed because it didn't know the codon ATG, since it's AA has only 1
codon. So we've gone back through the mapping functions and private
variables again. We now have set up the private variables to use all of
the AAs, which should now just have vectors of size 0 for unused AAs.
- added the split static table